Fix for primary keys with non-string datatypes
If there is a primaryKey defined, and that column has a non-string
datatype, you get a crash like:

```
in `parse': undefined method `gsub!' for 1.0:Float (NoMethodError)
```

The crash is caused by trying to parse a string as a number, but the
input value is already a numeric type.

This in turn is caused by calling column.validate twice on the same csv
cell. The first time through, the cells are validated and inflated, and
stored in the values array. Then when the primary keys are checked, we
try to re-validate the already inflated values.

The fix is to just use them as-is, dont re-validate.